Refactors Base.tsx into Base/*.tsx
